transform into system info only call
